Added multilingual capability.
Added in tests for the language meta field used by the MultiLanguage
plugin so the two can be used simultaneously.
The list of pages is then trimmed so that only pages of the same
language as the current page are offered. (The MultiLanguage plug-in
can be used to provide a link to the current page in different
languages.)